JAVAFX Scene connection test.
