+++ src/main/java/Main/Controllers/AdminController.java
@@ -1,0 +1,22 @@
+package Main.Controllers;
+
+import javafx.fxml.Initializable;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.input.MouseEvent;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class AdminController implements Initializable {
+
+    public Label username_label;
+
+    @Override
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+
+    }
+
+    public void refresh_db(MouseEvent mouseEvent) {
+    }
+}
