+++ src/main/java/Main/Main.java
@@ -20,6 +20,7 @@ public class Main extends Application {
         stage.setTitle("Login Screen");
         Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("scenes/login.fxml")));
         Scene scene = new Scene (root, 350, 500);
+        stage.setResizable(false);
         stage.setScene(scene);
         stage.show();
     }
