+++ src/main/resources/scenes/admin.fxml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ButtonBar?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Main.Controllers.AdminController">
+    <BorderPane prefHeight="400.0" prefWidth="600.0">
+        <bottom>
+          <ButtonBar prefHeight="56.0" prefWidth="600.0" BorderPane.alignment="CENTER">
+              <buttons>
+                  <ImageView id="refresh-bt" fitHeight="45.0" fitWidth="47.0" onMouseClicked="#refresh_db" pickOnBounds="true" preserveRatio="true">
+                      <Image url="@../Images/Refresh_button.png" />
+                  </ImageView>
+                  <Button mnemonicParsing="false" text="Button" />
+              </buttons>
+          </ButtonBar>
+        </bottom>
+        <right>
+            <Pane prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER">
+                <Label fx:id="username_label" alignment="TOP_LEFT" layoutX="14.0" layoutY="69.0" prefHeight="50.0" prefWidth="154.0" text="Username:">
+                    <font>
+                        <Font size="14.0" />
+                    </font>
+                </Label>
+                <Label layoutX="39.0" layoutY="22.0" text="Logged in user:">
+                    <font>
+                        <Font size="18.0" />
+                    </font>
+                </Label>
+            </Pane>
+        </right>
+        <left>
+          <TableView prefHeight="344.0" prefWidth="409.0" BorderPane.alignment="CENTER">
+              <columns>
+                  <TableColumn prefWidth="75.0" text="ID" />
+                  <TableColumn prefWidth="75.0" text="Username" />
+                  <TableColumn prefWidth="95.0" text="Password" />
+              </columns>
+          </TableView>
+        </left>
+    </BorderPane>
+</AnchorPane>
