+++ src/main/resources/scenes/login.fxml
@@ -5,6 +5,8 @@
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.effect.Glow?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
@@ -12,21 +14,21 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<GridPane alignment="center" hgap="10" prefHeight="501.0" prefWidth="352.0" vgap="10" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Main.Controller">
+<GridPane alignment="center" hgap="10.0" prefHeight="500.0" prefWidth="350.0" vgap="10.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Main.Controllers.Login_controller">
    <columnConstraints>
-      <ColumnConstraints halignment="CENTER" hgrow="NEVER" maxWidth="-Infinity" minWidth="-Infinity" />
+      <ColumnConstraints halignment="CENTER" hgrow="NEVER" maxWidth="1.7976931348623157E308" />
    </columnConstraints>
    <rowConstraints>
-      <RowConstraints valignment="CENTER" vgrow="NEVER" />
+      <RowConstraints maxHeight="1.7976931348623157E308" minHeight="-Infinity" valignment="CENTER" vgrow="NEVER" />
    </rowConstraints>
-   <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="500.0" prefWidth="350.0" style="-fx-background-color: #1c1449;" GridPane.halignment="CENTER" GridPane.hgrow="NEVER" GridPane.valignment="CENTER" GridPane.vgrow="NEVER">
-      <Button fx:id="login" layoutX="152.0" layoutY="375.0" mnemonicParsing="false" onAction="#log_in_function" text="Login" />
+   <AnchorPane prefHeight="500.0" prefWidth="350.0" style="-fx-background-color: #1c1449;" GridPane.halignment="CENTER" GridPane.valignment="CENTER">
+      <Button fx:id="login" layoutX="100.0" layoutY="418.0" mnemonicParsing="false" onAction="#log_in_function" prefHeight="25.0" prefWidth="150.0" text="Login" />
       <TextField fx:id="usernameField" layoutX="88.0" layoutY="227.0" promptText="Username">
          <font>
             <Font size="14.0" />
          </font>
       </TextField>
-      <Text fill="#e48787" layoutX="69.0" layoutY="202.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Login into your account:">
+      <Text fill="#889fe3" layoutX="69.0" layoutY="202.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Login into your account:">
          <font>
             <Font size="20.0" />
          </font>
@@ -39,6 +41,13 @@
             <Font size="14.0" />
          </font>
       </PasswordField>
-      <Label fx:id="label" alignment="CENTER" layoutX="59.0" layoutY="330.0" prefHeight="27.0" prefWidth="232.0" textAlignment="CENTER" textFill="#e48787" wrapText="true" />
+      <Label fx:id="label" alignment="CENTER" layoutX="59.0" layoutY="330.0" prefHeight="27.0" prefWidth="232.0" textAlignment="CENTER" textFill="#889fe3" wrapText="true">
+         <effect>
+            <Glow level="0.64" />
+         </effect></Label>
+      <ImageView blendMode="ADD" fitHeight="100.0" fitWidth="100.0" layoutX="125.0" layoutY="48.0" pickOnBounds="true"
+                 preserveRatio="true">
+         <Image url="@../Images/Avatar.png"/>
+      </ImageView>
    </AnchorPane>
 </GridPane>
